patch: extract Cellular2Settings to its own module
